Clear MPD state on non-0 exit from MPD song sensor
This is only half-arse general solution - while it catches and propagates non-0
exits for all sensors, not all sensors will actually propagate failures from
middles of their pipelines; and controller only handles the mpd case here.

The right thing to do seems to be to save the current timestamp value and
mark every key in the db with it, then only use values that are fresh-enough.
This needs a re-organization of how values are formatted by the sensors and
stored by the controller.

For error-handling we need to enable pipefail option on all sensor scripts,
but, how to parse and handle errors _well_? It seems that shell/awk are no
longer an advantage...